requirements.txt
acres==0.5.0
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.2.1
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
dotenv==0.9.9
etelemetry==0.3.1
fastapi==0.115.14
filelock==3.18.0
fitz==0.0.1.dev2
google-auth==2.40.3
google-genai==1.24.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
looseversion==1.3.0
lxml==6.0.0
MarkupSafe==3.0.2
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
numpy==2.3.1
packaging==25.0
pandas==2.3.0
pathlib==1.0.1
prov==2.1.1
puremagic==1.29
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pydot==4.0.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pyxnat==1.6.3
rdflib==7.1.4
requests==2.32.4
rsa==4.9.1
scipy==1.16.0
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
tenacity==8.5.0
traits==7.0.2
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
